+++ sdk/go/core/ydk/entity_lookup.go
@@ -17,8 +17,8 @@ func RegisterEntity(name string, entity_type reflect.Type) {
 func GetTopEntity(name string) (types.Entity, bool) {
 	_, ok := topEntityRegistry[name]
 	if !ok {
-                YLogError(fmt.Sprintf("Entity '%s' is not registered. Please import corresponding package to your application.", name))
-                return nil, ok
+		YLogError(fmt.Sprintf("Entity '%s' is not registered. Please import corresponding package to your application.", name))
+		return nil, ok
 	}
 	return reflect.New(topEntityRegistry[name]).Elem().Addr().Interface().(types.Entity), ok
 }